requirements.txt
astroid==2.4.2
autopep8==1.5.7
backports.entry-points-selectable==1.1.0
beautifulsoup4==4.9.3
bs4==0.0.1
certifi==2020.12.5
cfgv==3.3.1
chardet==3.0.4
click==7.1.2
cloudscraper==1.2.58
colorama==0.4.4
cssutils==2.3.0
dateparser==1.0.0
decorator==5.0.9
demjson==2.2.4
demjson3==3.0.5
distlib==0.3.3
filelock==3.3.1
flake8==4.0.1
Flask==2.0.2
gallery-dl==1.16.0
html5lib==1.1
identify==2.3.0
idna==2.10
isort==5.6.4
itsdangerous==2.0.1
Jinja2==3.0.2
joblib==1.1.0
lazy-object-proxy==1.4.3
LEPL==5.1.3
MarkupSafe==2.0.1
mccabe==0.6.1
mechanize==0.4.5
nodeenv==1.6.0
Pillow==8.2.0
platformdirs==2.4.0
pre-commit==2.15.0
psycopg2==2.8.6
psycopg2-binary==2.9.1
psycopg2-pool==1.1
py==1.10.0
pycodestyle==2.8.0
pycryptodome==3.10.1
pyflakes==2.4.0
pylint==2.6.0
pyparsing==2.4.7
PySocks==1.7.1
python-dateutil==2.8.1
python-magic==0.4.24
pytz==2021.1
PyYAML==6.0
redis==3.4.1
rb==1.9.0
regex==2021.8.3
requests==2.25.1
requests-toolbelt==0.9.1
retry==0.9.2
setproctitle==1.2.2
six==1.15.0
soupsieve==2.1
sqlparse==0.4.1
tabulate==0.8.7
text-unidecode==1.3
toml==0.10.2
tzlocal==2.1
ujson==4.2.0
urllib3==1.26.2
virtualenv==20.8.1
webencodings==0.5.1
websocket-client==1.1.0
Werkzeug==2.0.2
wrapt==1.12.1
yoyo-migrations==7.3.2
